+++ src/Datadog.Trace/Agent/NamedPipes/IDial.cs
@@ -1,11 +1,10 @@
-using System.IO;
 using System.IO.Pipes;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace Datadog.Trace.Agent.NamedPipes
 {
-    public interface IDial
+    internal interface IDial
     {
         Task<NamedPipeClientStream> DialAsync(TraceRequest request, CancellationToken cancellationToken);
     }
