+++ src/Datadog.Trace/Agent/NamedPipes/NamedPipesApiRequest.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Datadog.Trace.Agent.MessagePack;
+using Datadog.Trace.Agent.NamedPipes;
 
 namespace Datadog.Trace.Agent
 {
