+++ src/Datadog.Trace/Agent/NamedPipes/TraceRequest.cs
@@ -1,19 +1,27 @@
 using System.Collections.Generic;
 
-namespace Datadog.Trace.Agent
+namespace Datadog.Trace.Agent.NamedPipes
 {
     /// <summary>
     /// A fake HTTP class
     /// </summary>
     internal class TraceRequest
     {
+        /// <summary>
+        /// Gets or sets the path to use on the server.
+        /// Determines the schema version for the traces.
+        /// </summary>
         public string Path { get; set; } = "/v0.4/traces";
 
-        public string Host { get; set; } = "localhost:8126";
+        /// <summary>
+        /// Gets or sets the HTTP Version.
+        /// Version 1.0 of HTTP is "safe".
+        /// </summary>
+        public string Version { get; set; } = "1.0";
+
+        public string Host { get; set; } = "localhost";
 
         public string Method { get; set; } = "POST";
-
-        public string Version { get; set; } = "1.0";
 
         public List<KeyValuePair<string, string>> Headers { get; set; } = new List<KeyValuePair<string, string>>();
